+++ lib/swift_transitions.dart
@@ -15,6 +15,7 @@ export 'src/page/back_gesture.dart' show BackGestureRegion;
 export 'src/page/swift_page_route.dart' show SwiftPage, SwiftPageRoute;
 export 'src/page/swift_page_transition.dart' show SwiftPageTransition;
 export 'src/page/swift_page_transitions_builder.dart';
+export 'src/physics/curves.dart';
 export 'src/physics/springs.dart';
 export 'src/zoom/zoom_dismiss_physics.dart';
 export 'src/zoom/zoom_gestures.dart' show ZoomScrollController;